pyproject.toml
[tool.poetry]
name = "byteme"
version = "0.1.0"
description = "this is a music data app"
authors = ["Sophie Subair"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
logging = "^0.4.9.6"
pyiceberg = "^0.9.1"
pre-commit = "^4.2.0"
protobuf = "^6.31"
billboard-py = "^7.1.0"
pandas = "^2.3.1"
pyspark = "3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
